requirement.txt
fastapi
uvicorn[standard]
groq
python-dotenv
pydantic
requests
tenacity
google-auth-httplib2
google-api-python-client
google-auth-oauthlib
gspread